Bugfix in Program::createBasicBlock
It no longer crashes if the result of the split is empty. This can
happen, e.g., when you want to split a basic block, whose successor
address is bigger then the address of the end of its instruction. An
example of such a basic block is the one which is created when
converting ret + nop to IR.